Organise code ready to implement external file chooser
Instead of using a plugin embedded in Moodle, move the logic
for the file chooser into an external pluginservice (running
in its own container) and modify the plugin to load that.

This is half-way between the entirely-embedded plugin and an
LTI Tool Provider. It means it's possible to provide a better
file chooser UI within Moodle and makes it easier to develop
the Acropolis "proxy" layer, as it's isolated from Moodle.

Also reorganise the code to make it simpler to run docker-compose.

This provides the first steps to the implementation.