Add plumbing for pluginservice
Get basic skeleton of pluginservice web API working.

This will provide the endpoint for filechooser requests coming
from Moodle, and will display search results and topic data
from Acropolis. When an image is clicked, details of the file
will be sent back to a callback URL on the Moodle instance.